+++ lib/params/twoDdataparams.cpp
@@ -606,7 +606,7 @@ elementStartHandler(ExpatParseMgr* pm, int depth , std::string& tagString, const
 	}
 	//Parse the geometry node
 	else if (StrCmpNoCase(tagString, _geometryTag) == 0) {
-		double exts[6];
+		float exts[6];
 		GetBox()->GetExtents(exts);
 		while (*attrs) {
 			string attribName = *attrs;
@@ -615,11 +615,11 @@ elementStartHandler(ExpatParseMgr* pm, int depth , std::string& tagString, const
 			attrs++;
 			istringstream ist(value);
 			if (StrCmpNoCase(attribName, _twoDMinAttr) == 0) {
-				ist >> (double)exts[0];ist >> (double)exts[1];ist >> (double)exts[2];
+				ist >> exts[0];ist >> exts[1];ist >> exts[2];
 				GetBox()->SetExtents(exts);
 			}
 			else if (StrCmpNoCase(attribName, _twoDMaxAttr) == 0) {
-				ist >> (double)exts[3];ist >> (double)exts[4];ist >> (double)exts[5];
+				ist >> exts[3];ist >> exts[4];ist >> exts[5];
 				GetBox()->SetExtents(exts);
 			}
 			else if (StrCmpNoCase(attribName, _cursorCoordsAttr) == 0) {
